backend/requirements.txt
asgiref==3.7.2
Django==4.2.2
djangorestframework==3.14.0
pip==21.3.1
pytz==2023.3
setuptools==60.2.0
sqlparse==0.4.4
tzdata==2023.3
wheel==0.37.1
